hal: Add device and queue wait for idle, refactor errors into separate module
