+++ src/hal/src/adapter.rs
@@ -3,6 +3,7 @@
 //! Physical devices are the main entry point for opening a [Device](../struct.Device).
 
 use {format, memory, Backend, Gpu, Features, Limits};
+use error::DeviceCreationError;
 use queue::{Capability, QueueGroup};
 
 /// Scheduling hint for devices about the priority of a queue.  Values range from `0.0` (low) to
@@ -76,46 +77,6 @@ pub trait PhysicalDevice<B: Backend>: Sized {
 
     /// Returns the limits of this `Device`.
     fn get_limits(&self) -> Limits;
-}
-
-/// Device creation errors during `open`.
-#[derive(Fail, Debug, Clone, PartialEq, Eq)]
-pub enum DeviceCreationError {
-    /// Memory allocation on the host side failed.
-    /// This could be caused by a lack of memory.
-    #[fail(display = "Host memory allocation failed.")]
-    OutOfHostMemory,
-    /// Memory allocation on the device side failed.
-    /// This could be caused by a lack of memory.
-    #[fail(display = "Device memory allocation failed.")]
-    OutOfDeviceMemory,
-    /// Device initialization failed due to implementation specific errors.
-    #[fail(display = "Device initialization failed.")]
-    InitializationFailed,
-    /// At least one of the user requested extensions if not supported by the
-    /// physical device.
-    #[fail(display = "One or multiple extensions are not supported.")]
-    MissingExtension,
-    /// At least one of the user requested features if not supported by the
-    /// physical device.
-    ///
-    /// Use [`get_features`](trait.PhysicalDevice.html#tymethod.get_features)
-    /// for checking the supported features.
-    #[fail(display = "One or multiple features are not supported.")]
-    MissingFeature,
-    /// Too many logical devices have been created from this physical device.
-    ///
-    /// The implementation may only support one logical device for each physical
-    /// device or lacks resources to allocate a new device.
-    #[fail(display = "Too many device objects have been created.")]
-    TooManyObjects,
-    /// The logical or physical device are lost during the device creation
-    /// process.
-    ///
-    /// This may be caused by hardware failure, physical device removal,
-    /// power outage, etc.
-    #[fail(display = "Physical or logical device lost.")]
-    DeviceLost,
 }
 
 /// Information about a backend adapter.
